move bulk of code to a new lib
